Add preview popup when clicking gutter icon
Clicking an icon in the editor gutter immediately opens a new tab
to view/edit the corresponding image resource. Users have suggested
that it would be preferable to instead show a preview of the image
with the option of switching to a new tab to edit it.

This CL resolves the issue by adding a level of indirection to
GutterIconRenderer#getClickAction which generates a larger popup
preview of a gutter icon when the icon is clicked. Clicking the
preview opens the image resource in a new tab. To quickly open the
image resource in a new tab, one simply needs to double-click the
gutter icon.

Bug: 38346420
Test: Manually tested
Change-Id: Id6d1f04b25d4fc60066cba2fbfda4b2d97292c03