Clean GutterIconFactory#replaceResourceReferences
The current implementation of replaceResourceReferences duplicates
the logic for resolving a single resource reference which exists in
ResourceHelper#resolveStringValue. It also uses a hardcoded int for
the maximum number of indirections it will follow before assuming
a cyclic dependency, instead of the static constant
ResourceResolver.MAX_RESOURCE_INDIRECTION which has been defined
for this purpose.

This CL resolves these issues by replacing the duplicate logic in
replaceResourceReferences with a call to
ResourceHelper#resolveStringValue, which in turn makes calls to
methods in ResourceResolver that use the appropriate static
constant.

Test: Covered by new tests in GutterIconFactoryTest.java
Change-Id: I5545069ab00fd71051059b30e3d687fc2d6764b1