Use timers instead of focus listeners
Fix enteringTextInDefaultValueTextFieldUpdatesTableCell flakiness. The
test selects another cell after entering text in the Default Value text
field. The text field used a focus listener for notification of the user
finishing entering text. The code assumed the focus lost event would
come before the table selection changed event that overwrites the field
but that wasn't always the case.

Fix the timing problem by doing away with the focus listener and using a
timer instead. Every key press in the text field resets the timer. After
the user stops typing and 500 milliseconds have passed, the value in the
field is set in the table which writes to the resource file (which is
what the old focus listener did).

Bug: http://b/63177330
Test: Existing tests
Change-Id: Iab641601bef29b4e7b2641057e83997204c49065