Moving static utility methods to utility classes
Moves public static methods from NlModel to utility classes. The methods
do not use or depend on NlModel and they are shared with other classes
so they make more sense in separate utiltiy classes.
Not the methods are unit tested and a small fix was fixed in
NlComponentUtil.isDescendant

Also refactored groupSiblings to return Multimap instead of
Map<?,List<?>>

Test: New tests
Change-Id: Ib6df8d0166b9f1f96a76749bcf873581f2b20b8a