Implement contextual call graph search
Use the Cartesian product algorithm on the call graph
to find contextual call paths from source nodes to sink nodes.
The context considered includes formal parameters and the implicit
receiver parameter.

The heart of the algorithm is in CallGraph.kt; the remaining changes
are in support of the algorithm. See the changes to
testData/callgraph/CallGraph.java and CallGraphTest.kt for examples of
how this algorithm gives us quite a bit more precision.

Test: CallGraphTest, InterproceduralThreadAnnotationTest
Change-Id: I477ac96ca3c561b1328f30bc4b3a2a8f0ecb0c08