Remove trailing/leading "." from MethodModel Id
We were not checking if we needed a separator between class name and
name/signature. As a result, our Id could end up having a trailing
and/or leading "." character.

First, we shouldn't have an empty method Id. This CL changes
ArtTraceHandler to make sure we set methodName instead of className for
special nodes created to represent a thread. That should remove the
trailing "." characters from the Id.

Leading "." character was removing by simply checking the presence
of a class name.

Test: Updated affected tests
Bug: 63557433
Change-Id: Ie548c3a187a820fa25b6da8836068bf5a5f9ffe1