UI test for generating an APK debugging and profiling project
The APK debugging and profile project is generated from
a locally built APK. This test verifies that the paths
to the NDK and project source code are properly found
for the native library when the library's debugging
symbols are added.

The APK project is built from the San Angeles Google samples
project.

Test: Run this test
Change-Id: I3839a3de589da22f87265944d161ff8c129147d0