Fix NewProjectMultiModuleTest.createAllModuleWithIApp
This test is failing on Build Bot because the Instant App SDK is not
installed, and the check box to select an iapp is then not displayed.
We now mock the sdk location. We just need to build a new project.
The SDK is only needed for deploying on a phone/emulator.

Test: This is a test
Change-Id: I5f970220ebaeab425907301d257d25ebc138dd8d