Precompute overriding method search for call graph
When using the overriding method search provided by IntelliJ, the call
graph analysis could run out of memory and crash after an hour or two.
Now, when running on support-lib, the analysis completes in 8 minutes.

This also removes the dependency on IntelliJ, allowing the analysis to
run as a Lint check outside the IDE.

Overriding method search does not yet analyze byte code.

Test: CallGraphTest, InterproceduralThreadAnnotationTest
Change-Id: I1c67aa0f64fdee9051d9b0b0e6ae86aacacefc7f