[AVD Mgr] Change the warning about modifying SD card size
If the user changes the size of an AVD's SD card, the emulated card
is erased. We warn the user before doing this.

This changes how that warning is presented. Previously, if the user
modified the size field and then clicked Finish, the warning message
would be instantly removed and the field was not validated.

With this change, the validity is checked with every keystroke
and an error/warning message is displayed as needed.

Bug: 63115269

Test: Verified the SD card size field and other "Storage" fields.

Change-Id: Iab5e201c686d8bd7aa8412053fe7d0e27bb44adb