Call graph: handle methods invoking on new thread
To improve thread annotation checks, this change looks for thread
annotations on parameters, which can indicate that a lambda argument (say)
will run on a different thread.

The code churn in CallGraph.kt supports this by giving client
code access to the parameter contexts used to initialize a search for
paths. The underlying algorithm and functionality there has not changed.

Test: InterproceduralThreadAnnotationTest
Change-Id: I98ed746ce2cb5eea41aa547f52f6110a7c1ddc4c