Don't report XML resolve issues in test files
Currently we don't handle test resources at all and usually don't even
recognize them as resources files. But IDE can be tricked into treating
them as resources if a test manifest file is in the right place (no idea
what this is supposed to achieve). For now let's just stop reporting
resolve issues in the inspection, until we can make it correct.

Bug: 62463347
Test: TestResourcesTest
Change-Id: I11a14b4afa804fed41f60885d8649890fc8b1f95