Preserve CPU profiling configuration when re-entering stage
We were not preserving the CPU profiling configuration (except the
profiler type) used to capture when stop capturing. One consequence of
that, for instance, is we were sending the current selected profiling
configuration to the feature tracker. That is wrong because of the
following example scenario:

1) Start a capture using SAMPLED configuration
2) Change the configuration to INSTRUMENTED
3) Click on "Stop"

In the above scenario, we were incorrectly sending the INSTRUMENTED
configuration to the tracker, while SAMPLED should be sent instead.

In order to fix that, we need to store the configuration properties
(i.e. profiler type, mode, buffer size limit and sampling rate) used
when starting the capture, so we make sure to retrieve it when stopping
the capture. This way we make sure to pass the correct properties to the
feature tracker. That can also be convenient in the future in case we
need to display information of the profiling configuration in the UI.

Bug: 63839603
Test: Unit test
Change-Id: Icec6443a0cd7890328290bcbdb39ea72c36162c7