Remove dependency on AndroidSdkUtil from AdbService
Removing the dependency on AndroidSdkUtil is a pre-requisite for
eventually moving AdbService to a separate module.

AndroidSdkUtil was only used to identify the location of adb given
a project. This CL changes restartDdmlib to take in the path directly,
and the callers were updated to pass in the path to the adb.

Test: no functional changes, existing tests pass
Bug: b/63850984
Change-Id: I8a3eb0b5549f57de760a96fae65cc7e52e1f23d2