[AIA] Fix isIntantAppEnabled method
Instant App sdk used to be obtained either from an environment variable,
a system property or from the local android sdk repository. This should
not be done anymore (now that the SDK is already public) and it was
causing problems of instant apps being considered enabled when test sdk
environment variable was used.

This CL fixes this problem. Now the SDK is obtained only though the
SdkHandler APIs and checking if it's enabled checks if the sdk is
downloaded.

Test: existing tests should pass

Bug b/63829126

Change-Id: I009900eeb4af66d94f7f53a427df4871d1d29dad