Do not display negative values in CPU usage chart
We're using /proc/stat to obtain CPU time information and display it in
the CPU usage chart. The problem is one of the states included in
/proc/stat (http://man7.org/linux/man-pages/man5/proc.5.html) is iowait,
which is the time waiting for I/O to complete, and this state is not
reliable. iowait may actually decrease in certain conditions and that
would cause us to display negative app and system usage. We need to cap
these values to 0% to make sure they are not displayed to the user.

Bug: 63790627
Test: Unit tests
Change-Id: I35ed6c10de2ef4b547f964f1c6765ab5aa0f757e