[AVD Mgr] Add Emulator controls to select fast or slow boot
The Emulator is adding support to boot from a previously-saved
image snapshot. The default is to use this feature, which makes
the AVD boot much faster.

Provide a control that allows the user to force an AVD to do the
old-style cold boot.

This is a further modification to CL 2540919

Bug: 37851424

Test: Working on it...

Change-Id: Ic9d512844457e67cb084dd19540a2be5597a0f9a