Use trackCaptureTrace to track data related to CPU captures
The proto files corresponding to the tracking of CPU captures were
recently changed. This CL makes the appropriate changes on the client
side, making sure to pass all the information to the tracker using a
CpuCaptureMetadata object.

From now on, we use FeatureTracker#trackCaptureTrace to track actions
related to CPU captures and traces. trackTraceCpu was already deprecated
and was removed in this CL.

Test: updated tests that were still relying on trackTraceCpu. The tests
      for trackCaptureTrace were already added in a previous CL.
Bug: 63454214

Change-Id: Idf0881b5b26d878265d65c0dbf80dea3ec41136b