Use level2 dependency graph and level2 library in new PDS.
This change includes,
replace Variant with IdeVariant, replace BaseArtifact with IdeBaseArtifact,
replace Dependencies with IdeDependencies, replace Library with level2.Library,
replace MavenCoordinates with GradleCoordinate.

Major difference is that IdeDependencies returns flattened list
of Library. So there is no need to traverse dependencies recursively.

However, It doesn't reflect dependency graph.
To show the full graph dependency, PDS needs to do a separate
sync with full graph mode, which is out of the scope of this CL.
This CL is to make sure the current tests pass when we switch to level 2
dependency model in ide.

Test: Existing PS tests are passing with level 2 enabled.
Change-Id: I99fd628957c81a5fc1c3f1ae79046403f6c5243f