Disconnect Perfdproxy when ADB is gone.
Every time when Studio's ADB integration is disabled/enabled
(for example, starting/closing Android Device Monitor while
Studio is running), |bridgeChanged(bridge)| callback is
called twice. |bridge| is null on the first call, and not
null on the second one.

On the first call, |bridge| being null effectively tells
that the prior ADB becomes invalid, and the running perfd is
being killed.

On the second call, |bridge| is valid and
StudioProfilerDeviceManager will spawn perfd again. It
will find a new available local port and establish the
adb-forward link to the perfd running on the device. (When
ADB integration is disabled, we don't see any clients/
debuggable processes.)

Because the old PerfdProxy is targeted to the old local port
number, it cannot talk to the new Perfd. Therefore, killing
the old PerfdProxy cleans up the way for data store and the
new perfd to communicate through a new PerfdProxy.

Notes:

The device isn't disconnected during the process. This is
why existing logic cannot handle the case.

It is okay for old and new PerfdProxy to use the same
channel (device's serial) because their life spans don't
overlap.

Related but not helpful in fixing this bug:

AndroidEnableAdbServiceAction.isAdbServiceEnabled() returns
true when Studio's ADB integration is enabled; and false
when ADB integration is disabled. We could use this to guard
the calls from PerfdProxy to perfd to avoid the exceptions
reported by the bug, but leaving a dead perfd' PerfdProxy
running causes issues if ADB integration is back later.

Bug: 63122996

Test: Manual.
Change-Id: I5eab82f6b9b52f25a7b10f0db87b305ffb3fe041