Add image zooming on Layout Inspector Editor
Add the standard set of image zoom actions to Layout Inspector
preview panel so user can zoom in to see preview details.
This is a precursor to adding Pixel Perfect feature to Inspector.
We will follow up by adding ability to select an overlay image
to layer on top of the preview for Pixel Perfect.

The preview is rendered with the following conditions:
Initial render, fit the preview to the given viewport.
On resize, if viewport decreases in size, allow for zoom out.

Revert "Revert "Add image zooming on Layout Inspector Editor""

This reverts commit e80af10602855f2761296f408de4baa7e9430afe.

Test: Added LayoutInspectorZoomTest
Change-Id: I9f596a0d58ecb471026ec57d7c5d7568f9d2625f