Execute assemble task for tested target variant
When running tests in a test-only module, Studio was running
:test:assembleVariant but not assembling the tested target variant
(:testedProject:assembleTestedVariant). This caused changes in the
tested project not being updated when running tests.

This CL fixes this issue by analysing this case in GradleTaskFinder and
making sure that all the tested variants of the tested projects are
assembled.

Test: GradleTaskFinderTestOnlyModuleTest

Fixes bug b/63957818

Change-Id: Ida26ba9a4bb460269f35b6b98eeb06c33b20bb77