Remove ROUNDED_BORDER_ARC_SIZE from ThemeEditorConstants
SwatchComponent.java is the only class which uses this constant and
is the location of the comment which describes the purpose of
the constant. Removing ROUNDED_BORDER_ARC_SIZE and in-lining the
definition of ARC_SIZE in SwatchComponent will make it easier to migrate
com.android.tools.swing.ui to the adt-ui module because it removes one
of the dependencies SwatchComponent has on the android module.

Test: n/a, just removed a level of indirection
Bug: 63841225
Change-Id: I7cb03cb004657dc274bbf9968a4695489ae57e5d