[AIA] Consumes post build model for InstantApps bundle
After CL ag/2481300 introducing a new post build model for Instant Apps,
the pre build model started returning the wrong value for the bundle
output.

This CL syncs the new model after building and before running and uses
it to get the right output for instant app modules.

TODO: since the plugin is returning the wrong value in pre build model,
old versions of Studio will be broken if run against new versions of the
plugin. We probably want to fix the plugin to return the right value in
pre buld models

Test: unit tests included

Partially fixes bug b/64137115

Change-Id: Icebfd6226762148ff89b3c4ede0715829ad0bd8a