[Bug: 38264918] Update list of available Android SDK before trying to setup
IntallPlatformQuickFix Installs an Android SDK platform and then tries to
set it up by calling AndroidSdkSetup.setUpSdk. This functions looks for the
list of available sdks by calling findSuitableAndroidSdk but since this sdk
was just installed, it is not in the ProjectJdkTable yet, causing
findSuitableAndroidSdk to return null.

This CL causes an update to ProjectJdkTable before calling setUpSdk.

Test: Modified testExecute to make sure the list of sdk's was updated.
Change-Id: Ifbb3166ff31aee52aa4ca69b6bc6d7a90901ce67