Add @NotNull annotation to myCaptureState and initialize it
This field should be not be null and it is not initialized in the
constructor, but in a method (updateCaptureState()) called by it.

Annotating it with @NotNull would result in a inspection warning,
because it was not initialized when declared nor in the constructor.
We initialize it with IDLE state to ensure the annotation is in place
and there are no warnings.

Test: N/A
Change-Id: I95b881666280afb4a41fb51834ad3f89981a9339