Add test for public methods of BuildActions
Using Guava classes in non private methods of a BuildAction can cause
class loader issues due to different versions of the library: the
BuildAction code is run inside Gradle, with Gradle classpath, which
contains Guava in it; when back to Studio, the methods are run with
Studio classpath, which might contain a different version of Guava
appearing first in it and then it can cause casting problems for
example.

This CL adds a test that verifies that no Guava classes are used in non
private methods of the existing BuildActions.

Test: PublicMethodsBuildActionTest

Fixes bug b/63929984

Change-Id: If92996ebb5e2dac3a9855ac895853e48e6021305