[AIA] Fix debug when application dependency is used
Application dependency of base module in another means that the package
name should be obtained from the other module. Thus, applicationId is
only known after build.

Debugging an Instant App was failing because the process name (derived
from the applicationId) was wrong and then the debugger was not attached
correctly.

This CL fixes this issue by consuming the post build model for instant
apps and querying the correct applicationId from there.

Test: GradleApplicationIdProviderTest.testGetPackageNameForInstantApps

Fix bug b/62615782

Change-Id: I9d1c75f12906d1b41b93e1b9a3e714e0b9cd7dd6