Create project-system and project-system-gradle modules.
Create the AndroidProjectSystem interface, the
com.android.project.projectsystem extension point, and the Gradle
implementation of it.

Refactor AnalyzeApkAction to make use of the new abstraction rather than
invoking gradle-specific code directly.

Change ApkParser to have it recieve the aapt path externally rather than
have it compute it in a gradle-specific way, and refactor callers to
supply that path.

Bug: 64250896
Test: Manually opened the APK analyzer in both a gradle and a blaze project.

Change-Id: I9faac3de4288af7bbe3d093cbb6d1ce7a0cb68ba