Remove Dynamic Wizard: Refactor TemplateTest
This is one of many CL's to remove the Dynamic Wizard.

This CL introduces clearer distinction between what is project state
data and module state data. It also helps with the future replacement of
NewProjectWizardState by the usage of TemplateValueInjector.
- New NewProjectWizardState no longer extends TemplateWizardState and
instead has a new TemplateWizardState myModuleState (In theory a project
may have more that one module, so technically this should be a list, but
that is out of the scope of this work).
- TemplateTest variables with names like value/values/state were renamed
with more meaningful names.

Test: N/A removing dead code
Change-Id: Ib4076a75d9b430db36c669b30259bfd11f4808ef