Turn off IR if hasCode is false
Disable Instant Run for the session if application's manifest has
hasCode set to false (default true). This means the app does not have
any Java code and thus is not applicable for IR.

Note: The final manifest in the APK might have hasCode="true" this is
due to a previous work around added in b/37132271.

Fixes: b/37133532
Test: added InstantRunGradleUtilsTest
Change-Id: I10aee99ec41857e372b7be0f6d9ab5604a22db18