Fix NPE when closing profilers with live allocation enabled
We are setting the device to null when closing/stopping the profilers.
When the current device is selected to null, we call stopProfiling in
all the profilers. The issue was memory profiler checks if live
allocation feature is enabled when stopping profiling, so it can suspend
it. The check was trying to verify the API level of the device without
checking for its nullability first.

Bug: 63745419
Test: added unit test
Change-Id: Iafcab71ec4a541652a66ca1b869e72f2cb16474b