Ensure Webp is registered before attempting to read a .webp file
The DeviceArtDescriptor test reads files present in the
device-art-resources folder. Some new artwork added there is in the webp
format, which requires a webp reader spi to be registered before
javax.image can read it. The DeviceArtDescriptorTest relied on some
other test registering the webp spi, and was therefore failing when run
by itself. This CL registers the web spi inside the test.

Test: updated
Change-Id: I4ce5348243efdcc8f7c3f9cfed950bb6671c76ca