[Bug: 37141013] Replace TemplateProjectSettingsGroup with Android version
TemplateProjectSettingsGroup is replaced by
AndroidTemplateProjectSettingsGroup only in "WelcomeScreen.Configure.IDEA"
but this group was never changed in the ActionManager, causing that searches
from "Find Action..." menu return TemplateProjectSettingsGroup and no
AndroidTemplateProjectSettingsGroup.

This CL also replaces TemplateProjectSettingsGroup in the ActionManager.

Test: Added a couple of tests for GradleSpecificInitializer that verify an
intance of AndroidTemplateProjectSettingsGroup is being used on the welcome
dialog and in the ActionManager. At this moment tests can be run on Idea
but not from bazel and are disabled (see ag/2695439).

Bug: 37141013

Change-Id: Ic092bbe0d529ed3782b4a393ffba73332f87c404