Remove Dynamic Wizard: Remove TemplateMetadata@ATTR_RELATIVE_PACKAGE
This is one of many CL's to remove the Dynamic Wizard.

ATTR_RELATIVE_PACKAGE is only set by
TemplateWizardState@populateRelativePackage() and this method is only
called from TemplateTest with a null parameter, meaning that
ATTR_RELATIVE_PACKAGE is always be set to "".
The functionality previously implemented on the Wizard template by
ATTR_RELATIVE_PACKAGE is now implemented by ManifestMerger2@extractFqcns
so this attribute is no longer needed.

Test: N/A removing dead code
Change-Id: I67e3d3665f8be8ad022630bb730c25782f006b21