Refactor SimpleperfTraceParser to cleanup and simplify code
Splitted the samples per thread before parsing them. This way, we don't
need to keep a map of last call chains and last visited nodes, which now
can be passed as arguments to the methods called by the parser.

This is going to be particularly useful when we update the version of
simpleperf we're using, as they recently added support to "scheduled out"
events. Parsing samples per thread will make it easier for us to set a
more accurate end timestamp to thread sample nodes.

Bug: 37294023
Test: Refactor only. SimpleperfTraceParserTest still pass.
Change-Id: I961d7e77038a60a6f4fa36fa910bedbdb0b033dc