Ensure Webp is registered for DeviceArtPainterTest
Tests in DeviceArtPainterTest attempt to read .webp images,
which requires a webp reader spi to be registered. Other tests
call the method which ensures the webp reader spi has been registered,
so this issue is only noticeable when DeviceArtPainterTest is
run by itself.

This CL resolves the issue by calling
WebpMetadata#ensureWebpRegistered in DeviceArtPainterTest in a
@BeforeClass setup method.

See ag/2659903 for a similar CL.

Test: DeviceArtDescriptorTest now passes when run by itself
Change-Id: Ib74c54a6455307f1667ffbeb33c317c87c238404