Moved live allocation tracking logic to MemoryProfiler
Before, the logic lives in MemoryMonitor, which can potentially
be bypassed if the user quickly enters the MemoryProfilerStage,
resulting in MemoryMonitor being destroyed and missing the agent
status change signal. By moving to MemoryProfiler, the signal
always gets picked up and avoids the problematic case.

Also rearranged the stopProfiling logic in setDevice so that
it is called before the device is changed. This allows tracking
suspension to work correctly, as it requires to retrieve the
device version during stopProfiling.

Bug: 64490742
Test: updated unit test
Change-Id: I21d6ee85bc51016540c78871d29d2a05821c008d