Fix threading issue in ModelWizardTest tests
TL;DR - BatchInvoker must be created on the same thread as its processing
strategy runs on to avoid a potential race condition.

ModelWizardTest creates wizards on the main thread, but their underlying
BatchInvokers (which is used for processing its properties like canGoNext,
canGoBack, etc.) were left to their default configuration, which runs on
the UI thread. This could cause some instances where a wizard
test might attempt to respond to a property changing on the main thread
while the UI thread happened to be in the middle of processing.

In production code, this is not an issue because all of our wizards are
created on the UI thread and processed on the UI thread. In test code,
we can solve this by setting BatchInvoker to use a test-only strategy
that runs on the main thread.

All other unit tests are already doing this except for this one, which was
written before the feature to globally override the default BatchInvoker
strategy was implemented, which is why it was missed.

Bug: 64293196
Change-Id: I3058bc7546b9bd78db35d316d24cbbc1d37c1686
Test: Only test code updated in this CL. Runs same as before but is now
thread safe.