[Bug: 64035630] Quickfix to add Google repository in build file...
... for missing dependencies.

When a com.android.support dependency is missing, a quickfix to install a
Google support repository is shown, this can also be solved by adding Google
Maven repository.

This cl checks if Google maven repository is already present and if not,
adds a AddGoolgeMavenRepositoryHyperlink quickfix for missing dependencies
that start with com.android.support.

AddGoolgeMavenRepositoryHyperlink was modified to add google repository
in the repositories for the buildscript and the repositories.

Test: Adapt tests for hyperlink and add new tests for new functions. Also
made changes to tests in UnresolvedDependenciesReporterIntegrationTest to
handle the cases when Google repository is used already.

Screenshot: https://screenshot.googleplex.com/5pXP1wSAuJ8

Change-Id: Iecf85e6a7b6a68eccecc681e8d23eb05228632aa