Remove Dynamic Wizard: Remove NewProjectWizardState.updateDependencies
This is one of many CL's to remove the Dynamic Wizard.

NewProjectWizardState.updateDependencies was needed before we had proper
merge of gradle files. The update of the dependencies is now done during
the project generation, and this is no longer needed. This will also
make sure that the code generated by the wizard and the tested code are
more similar.

Test: Run all TemplateTests
Change-Id: Ia48086ab3d1e5361c5fbed75f7e899da728b6330