Show manifest merging errors when adding templates
When adding a new template to a project where the template adds
content to the manifest (such as an activity registration), the IDE
runs the manifest merger combining the current manifest with the
manifest fragments from the template.

If the manifest merger rejects the merge, the template machinery will
just concatenate the old manifest and the new manifest (into what is
now an invalid XML document) and tell the user to merge the files
themselves. However, it's not obvious to the user *why* the merge
failed.

This CL improves this situation by reading out the manifest merger's
error messages (which is available in the activity log) and showing
these in an error log. A subsequent CL will also update the merging
process to pass a feature flag to the manifest merger to be more
accepting of some types of errors, such as two conflicting attribute
values.

Bug: 64552606
Test: Manual
Change-Id: Ib84bbf9745ba1cf33e3137a039ee9e17236cbac5
(cherry picked from commit 06e0f8938be71266ed56e52fae033424ce70fba6)