Update fix for checking project path on open.
The extra rules under withAllRules is not applicable to selecting existing
projects. Create a ProjectImportPathValidator class to contain logic for
validating a project file (really a dir).

It's a wrapper around PathValidator with commonRules for now but this
allows us to change the internal implementaion and edit rules as needed
without changing code that uses it.

Added the same check for selected project dir to
AndroidImportProjectAction as it should have the same requirements as
opening a project.

Also renamed AndroidOpenFileAction to AndroidOpenProjectAction to better
reflect its purpose.

Original commit: If188687cf2d2d1105b32fb55b8e4c60c132f9312
This reverts commit a377e2dfa28c07f63b6c430d809168f995d07857.

Fixes: b/64584325
Test: updated AndroidOpenProjectActionTest
Change-Id: I814a3f765afb601fe858554d21c7fc5dd520464f