Recover from NPE thrown by AndroidArtifact#getOutputs.
When migrating a project from older versions of the plugin
to 3.0.0-beta1, AndroidArtifact#getOutputs throws a NPE
making Gradle Sync fail.

The fix is to simply ignore 'getOutputs'. For plugin 3.0
we no longer care about this method, since we get the
build output when running the application. This method
is only useful when syncing pre-3.0 plugins.

Bug: 64286227

Test: JUnit test
Change-Id: Ic5bd21692b3d807858e44dd5aba05354f8358acf