Support attributes for custom NestedScrollViews.
Bug: b/64288820

Move the logic for adding all ScrollView attributes to
NestedScrollView into getSuperClass. This will add support for
fillViewport on classes derived from NestedScrollView.

The NestedScrollView itself is based on RelativeLayout and the v4
support library does not support custom attribute definitions.
So we need to hardcode for these support classes.

Test: Added.
Change-Id: I10bf413a39581f994dfb93e811e26458a0ddeedf