Refrain from closing the preview unless not needed.
We used to close the preview if any layout file was closed.
If 2 layout files are open with an open preview, the
preview should NOT be closed if the next active editor is
the other layout.

Bug: b/64288544

Test: Added ui regression test.
Change-Id: Ic7409110fa8d764c28fed34d6d3e66f0bc6ebd01
(cherry picked from commit 044c289b1e033578dda604bd1be4fba6db6f54f1)