[AIA] Add ProvisionBeforeRunTask to run configs after sync
ProvisionBeforeRunTask is already being automatically added to run
configurations newly created. However, when opening a project that does
not contain the task yet, it was not being added to old run configs.

This CL fixes this issue by adding the before run task to android
configurations in PostSyncProjectSetup.

Test: PostSyncProjectSetupTest.testProvisionBeforeRunTaskIsAdded
Fix bug b/64572730

Change-Id: I3ab23e971ee9ff61820b1a09f3d4e083978d456b