Attach IDs to providers of AndroidProjectSystem extensions
The IDs are provided by a new read-only attribute on AndroidProjectSystem.
Note that providing the IDs via XML was also considered, but this was
rejected since that would create duplication between the XML and any
code that explicitly attaches IDs to the project.

Adds a check for the default project system.

Test: Confirmed that the APK analyzer still launches
Bug: 64975051
Change-Id: I58ffc64f9f569593147b4e442b8df53881adc185