Ensure no indexing during sync & build (behind the flag).
This is the implementation which achieves the objective of preventing
indexing during sync and build.

Implied changes:
(1) We are saving time by not starting and then cancelling indexing
tasks when a new project root change event comes in. Instead, all such
tasks are queued up and then just one of them is executed as necessary,
because others are eliminated by IDEA according to its 'equivalence
object' principle for dumb mode tasks. There may be some cancellations
if the tasks of the same kind were not added sequentially.
(2) In the UI, there is another background task, which says that
indexing is blocked during gradle invocation.

Test: To be followed up.
Change-Id: Ie49fcbd84220f9f5cbd57575ed7018f3a4c8e0b0