Revert "Remove logcat from Run & Debug toolwindows"
This reverts commit 720694affb7ad59d89ceca9336ce6757e247827b.

After the revert, the code was slightly refactored with better
encapsulation and improved multi-threading support. The change
also contains changes related to enable unit testing, as well
as a new StudioFlags to enable the feature (disabled for now).

An upcoming CL will add a UI option to toggle the StudioFlag.

Bug: 37137285
Test: Unit Tests included
Change-Id: I6633cf77aef1201621c9c617feec0fe7b2960a14