Turn off delayed initialization for non-gradle projects
A better fix for this problem would be to move this feature to the build
system layer - BuildSystemService. What is required is the ability to
run a lambda once Gradle sync + source generation is complete for the
first time. The correct fix will be performed as part of b/65213155.

Cherry picked from If0d712b66e7f5290165218de1687ce12957958d3 and
resolved conflicts in c.a.t.idea/common/editor/NlEditorPanel.java

Test: existing tests pass
Bugs: 64889747, 64897027
Change-Id: If0d712b66e7f5290165218de1687ce12957958d3