Adds test for testNewProjectWithGoogleMapsWearActivityWithKotlin.
testNewProjectWithBlankWearActivityWithKotlin already exists.

Important Note: This CL was already +2 and submitted (original
CL/63610410) but
reverted because the test CL associated with it required the sandboxed
Android Studio test instance to have the latest version of the
support library (which it did not). It will have it in there before
submit this time.

Bug: 63610410
Test: This is a test for a template CL, so it is the test itself.
Change-Id: Ie33c3efb14e7b45afaf0f184cd3ac58a14d1a187
(cherry picked from commit 9b5693a95f00347524ef4449b8aec2604cd3c670)