Guarantee uniqueness of AndroidProjectSystem
The previous implementation of ProjectSystemComponent contained
a race condition that would have allowed multiple instances
of AndroidProjectSystem to be created for the same project.
This wasn't a problem since all instances are currently
interchangeable, but we want to offer a uniqueness invariant
for some planned future changes.

Test cases have been written but due to problems getting the
test module working in blaze, I've extracted them to a separate
commit in order to get this in.

Bug: 64976512
Test: Tests in a follow up
Change-Id: I0b5f0838eca28bab46dc5d846e4fed33513ae2ff