[UI Tests] NewInstantAppModuleTest failing
app_name String was not being created, when adding a new instant app to
a "non instant app" project and the "Instant App Wrapper" gradle file
was not created too.
The copy of the Strings and creating of the "Instant App Wrapper" is
controlled by ATTR_HAS_INSTANT_APP_WRAPPER and it needs to be true when
a new project doesn't have a "Instant App Base Module" yet.

Bug: 65237037
Test: NewInstantAppModuleTest pass
Change-Id: I0f2bd2cc81a5bacd283c4fd88c23441cfe1518a9