Use the declarative API for injecting Room SQL
We used to implement ConcatanationAwareInjector, which seems like part
of Java langauge implementation, not something for defining injections.
Now the strings that we know to be SQL are defined in XML, as if they
were defined in the settings UI. This also means that Kotlin editor can
recognize them, but unfortunately it doesn't currently work for
annotations, like @Query.

Test: RoomSqlLanguageInjectionTest
Change-Id: I0da4fa52f7136aca6800faf8fdaa92d083a5586d