Remove redundant UI update from GradleBuildSystemService#syncProject
The explicit call to BuildVariantView#projectImportStarted
in GradleBuildSystemService is redundant, as there is logic for
doing this in GradleSyncState. This CL removes the redundant call
and makes sure the equivalent logic is executed by setting the
myNewOrImportedProject field of the Gradle sync request appropriately.

Removing this redundancy also makes it so that
GradleBuildSystemService#requestSync doesn't need to be run on the EDT,
since the UI updates triggered by GradleSyncState are wrapped in an
invokeLater() call.

Bug: 65248191
Test: manually confirmed BuildVariantView doesn't display "Loading..."
      when syncing for a new project, but does display the message
      when importing an existing project as a module.

Change-Id: I55b386b1f407914c29ed33422b6fa832eb96b530