Fixed the weird quick definition for layout file
Avoid using layout editor to display the quick
definition.

The quick definition uses the first preferred editor
to display, but layout editor is not a good editor
to display the definition.
We still want to keep Layout Editor as default editor
of layout file, so only move back the priority when
policy is called by quick definition component.

Test: Manually test
Bug: 37050828
Change-Id: I7632a32c1282aba503f378e7225d2799c5d2fbdc