Flip the flag for indexing-aware gradle execution to true by default.
This is disabled by default in 3.0, however in order to get more
useful feedback from early 3.1 canaries, it's worth enabling it
on master. It still remains behind the flag so early adopters will
be able to disable the functionality if necessary.

This CL also ensures that indexing suspension is omitted in unittest
mode. See the comment in the code for more details.

Test: Existing. The entire UI testsuite will cover the new behavior now.
Change-Id: I67e6d10a9bae47259700bbb8e57a0678bcf8c280