Granular control over indexing during sync before project is initialised.
It appears that Gradle Sync can by design be invoked when project is
not yet initialised. This is a special case which needs to be handled
separately in the context of indexing control, otherwise it's easy to
run into a deadlock because IDEA would expect 'initial refresh' to
happen in order for the project to be initialised, hence the dumb queue
can't be put on hold by IndexingSuspender.

Instead, we employ more granular control in that case, and invoke
IndexingSuspender when project is initialised but before project setup.
This solves the workflow problem while still retaining concurrency
requirements.

This CL also improves error reporting and documents the expected usages.

Test: UI test suite covers this and this CL was inspired by the recent
UI test failures indicating the project initialisation corner case.

Change-Id: I663b02bb9b6fbff188eba01f2fcf1ca4e4823fb6