Memory leak: DependencyGraph leaks NlComponents
DependencyGraph.ourCache acted as a cache of DependencyGraph objects
for NlComponent instances. The map had weak keys (NlComponent), but held
strong values (DependencyGraph). However, DependencyGraph's object network
would eventually have a reference back to the NlComponent that was used as
a key in the weak map. According to WeakHashMap documentation, this will
create strong reference to the key and the entry will never be removed, even
if the reference from the key is the last reference to such NlComponent.

Fix: Removed ourCache from DependencyGraph and moved caching of the
DependencyGraph to NlComponent objects. This way, if the NlComponent is no
longer reachable, DependencyGraph references will not hold the object from
being garbage-collected.

Test: None
Change-Id: I0e89a093312dec87c4e3da9f3fa3ca405f45a4b1